// app/types/helpers.d.ts
export interface ObjectOf<T> {[key: string]: T}

export enum Food {
  alcohol = 'alcohol',
  burger = 'burger',
  chinese = 'chinese',
  fusion = 'fusion',
  healthy = 'healthy',
  italian = 'italian',
  pizza = 'pizza',
  seafood = 'seafood',
  steakhouse = 'steakhouse',
  sushi = 'sushi',
  mexican = 'mexican',
  thai = 'thai',
}

export enum Goal {
  serious = 'serious',
  exclusive = 'exclusive',
  casual = 'casual',
  unsure = 'unsure',
}